feat(php): use isolated php for non-interactive shells
Removed the composer and php aliases and instead added a small script to call the correct php binary.

This script is placed in the macos bin folder and we add the folder to the PATH in the .zprofile within the install-mac.sh.
Adding it after brew registers its paths, allows our script to be priortized instead of the currently linked php binary.
Because the .zprofile file is used to setup environment variables and it being called for non-interactive sessions too, calls to php in JS scripts for example will now correctly use the isolated php version via valet.