fix(macos): Link GNU patch for composer patches
The composer-patch util otherwise fails to apply patches.
MacOS ships with its own version of patch and this seems to need interaction, which then causes a timeout.

We now instead use GNU patch and also load it into our PATH.
This should work everywhere and be closer to what you expect when searching for usage online.